bankId Integer to Long

+++ core/src/main/java/uz/maroqand/ecology/core/dto/api/PaymentResponseData.java
@@ -10,10 +10,10 @@ import lombok.Data;
 @Data
 public class PaymentResponseData {
 
-    private Integer id;
+    private Long id;
     private Integer eco_id;
 
-    public PaymentResponseData(Integer id, Integer eco_id){
+    public PaymentResponseData(Long id, Integer eco_id){
         this.id = id;
         this.eco_id = eco_id;
     }
